move ddtrace.settings module into a folder
