+++ ddtrace/settings/__init__.py
@@ -1,0 +1,18 @@
+from .config import Config
+from .exceptions import ConfigException
+from .hooks import Hooks
+from .http import HttpConfig
+from .integration import IntegrationConfig, IntegrationConfigItem
+
+__all__ = [
+    'config',
+    'Config',
+    'ConfigException',
+    'Hooks',
+    'HttpConfig',
+    'IntegrationConfig',
+    'IntegrationConfigItem',
+]
+
+# Configure our global configuration object
+config = Config()
