+++ ddtrace/settings/exceptions.py
@@ -1,0 +1,5 @@
+class ConfigException(Exception):
+    """Configuration exception when an integration that is not available
+    is called in the `Config` object.
+    """
+    pass
